% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openprescribingRplots.R
\docType{package}
\name{openprescribingRplots}
\alias{openprescribingRplots}
\alias{openprescribingRplots-package}
\title{OpenprescribingRplot: A small package for creating plots in R from openprescribing data.}
\description{
OpenprescribingRplot: A small package for creating plots in R from openprescribing data.
}
\section{openprescribingRplot Functions}{


* plot2017(chemical_section_or_presentation_code = NULL)
}

