% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2017.R
\name{plot2017}
\alias{plot2017}
\title{Plot a leaflet map of cost per person, per CCG, per month of a BNF section, chemical or presentations.}
\usage{
plot2017(chemical_section_or_presentation_code)
}
\arguments{
\item{chemical_section_or_presentation_code}{An ID unique to BNF sections, chemicals or presentations.}
}
\value{
Returns a leaflet plot (html widget) of cost per person, per CCG, per month of the input BNF section, drug or chemcial.
}
\description{
Plot a leaflet map of cost per person, per CCG, per month of a BNF section, chemical or presentations.
}
\examples{
A plot of BNF section 7.4.5 spending = plot2017("7.4.5")
}
