% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2017items.R, R/plot2017quantityperitem.R
\name{plot2017items}
\alias{plot2017items}
\alias{plot2017items}
\title{Plot a leaflet map of total items, per CCG, per month of a BNF section, chemical or presentations.}
\usage{
plot2017items(argument)

plot2017items(argument)
}
\arguments{
\item{argument}{An ID unique to BNF sections, chemicals or presentations.}

\item{argument}{An ID unique to BNF sections, chemicals or presentations.}
}
\value{
Returns a leaflet plot (html widget) of total items, per CCG, per month of a BNF section, chemical or presentations.

Returns a leaflet plot (html widget) of quantity per item, per CCG, per month of the input BNF section, drug or chemcial.
}
\description{
Plot a leaflet map of total items, per CCG, per month of a BNF section, chemical or presentations.

Plot a leaflet map of quantity per item, per CCG, per month of a BNF section, chemical or presentations.
}
\examples{
plot2017items("7.4.5")
plot2017quantityperitem("7.4.5")
}
