% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2017listsize.R
\name{plot2017listsize}
\alias{plot2017listsize}
\title{Plot a leaflet map of list size per CCG, per month.}
\usage{
plot2017listsize()
}
\value{
Returns a leaflet plot (html widget) of total list size, per CCG, per month.
}
\description{
Plot a leaflet map of list size per CCG, per month.
}
\examples{
plot2017listsize()
}
