% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2017quantity.R
\name{plot2017quantity}
\alias{plot2017quantity}
\title{Plot a leaflet map of total item quantity, per CCG, per month of a BNF section, chemical or presentations.}
\usage{
plot2017quantity(argument)
}
\arguments{
\item{argument}{An ID unique to BNF sections, chemicals or presentations.}
}
\value{
Returns a leaflet plot (html widget) of total item quantity, per CCG, per month of a BNF section, chemical or presentations.
}
\description{
Plot a leaflet map of total item quantity, per CCG, per month of a BNF section, chemical or presentations.
}
\examples{
plot2017quantity("7.4.5")
}
