% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2017quantityperitem.R
\name{plot2017quantityperitem}
\alias{plot2017quantityperitem}
\title{Plot a leaflet map of quantity per item, per CCG, per month of a BNF section, chemical or presentations.}
\usage{
plot2017quantityperitem(argument)
}
\arguments{
\item{argument}{An ID unique to BNF sections, chemicals or presentations.}
}
\value{
Returns a leaflet plot (html widget) of quantity per item, per CCG, per month of the input BNF section, drug or chemcial.
}
\description{
Plot a leaflet map of quantity per item, per CCG, per month of a BNF section, chemical or presentations.
}
\examples{
plot2017quantityperitem("7.4.5")
}
