% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2017total.R
\name{plot2017total}
\alias{plot2017total}
\title{Plot a leaflet map of total cost per CCG, per month of a BNF section, chemical or presentations.}
\usage{
plot2017total(argument)
}
\arguments{
\item{argument}{An ID unique to BNF sections, chemicals or presentations.}
}
\value{
Returns a leaflet plot (html widget) of total cost, per CCG, per month of the input BNF section, drug or chemcial.
}
\description{
Plot a leaflet map of total cost per CCG, per month of a BNF section, chemical or presentations.
}
\examples{
plot2017total("7.4.5")
}
