% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openprescribingRplots.R
\docType{package}
\name{openprescribingRplots}
\alias{openprescribingRplots}
\alias{openprescribingRplots-package}
\title{OpenprescribingRplots: A small package for creating plots in R from openprescribing data.}
\description{
OpenprescribingRplots: A small package for creating plots in R from openprescribing data.
}
\section{openprescribingRplots Functions}{


* plot2017perperson(argument = NULL)
* plot2017total(argument = NULL)
* plot2017listsize()
* plot2017items(argument = NULL)
* plot2017quantity(argument = NULL)
* plot2017quantityperitem(argument = NULL)
}

