% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2017items.R
\name{plot2017items}
\alias{plot2017items}
\title{Plot a leaflet map of total items, per CCG, per month of a BNF section, chemical or presentations.}
\usage{
plot2017items(argument)
}
\arguments{
\item{argument}{An ID unique to BNF sections, chemicals or presentations.}
}
\value{
Returns a leaflet plot (html widget) of total items, per CCG, per month of a BNF section, chemical or presentations.
}
\description{
Plot a leaflet map of total items, per CCG, per month of a BNF section, chemical or presentations.
}
\examples{
plot2017items("7.4.5")
}
